% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapshiny_auxi.R
\name{addHeatmapAnnotation}
\alias{addHeatmapAnnotation}
\title{Prepare data for heatmap annotations}
\usage{
addHeatmapAnnotation(x, column = TRUE, var.name = "")
}
\arguments{
\item{x}{a vector or matrix used for annotation. nrow of matrix should equal ncol/nrow of heatmap matrix}

\item{column}{whether the annotations are for columns}

\item{var.name}{variable name}
}
\description{
Prepare data for heatmap annotations
}
