% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapshiny_auxi.R
\name{addHeatmapAnnotation_plot}
\alias{addHeatmapAnnotation_plot}
\title{plot heatmap annotations}
\usage{
addHeatmapAnnotation_plot(object, ...)
}
\arguments{
\item{object}{an object returned by \link{\code{addHeatmapAnnotation}}}

\item{...}{other parameters passed to plot, usually xlim or ylim}
}
\description{
plot heatmap annotations
}
