% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapshiny_auxi.R
\name{adist}
\alias{adist}
\title{Prepare data for heatmap annotations}
\usage{
adist(x, method = "pearson")
}
\arguments{
\item{x}{a matrix, the distance between rows are calculated}

\item{method}{method used for distance calculation}
}
\description{
Prepare data for heatmap annotations
}
