% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_figureAttr4.R
\name{attr4selector_module}
\alias{attr4selector_module}
\title{Utility - extended figure control shiny module}
\usage{
attr4selector_module(input, output, session,
  reactive_meta = reactive(NULL), reactive_expr = reactive(NULL),
  reactive_triset = reactive(NULL))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_meta}{reactive meta info, usually phenotype data or feature data of ExpressoinSet}

\item{reactive_expr}{expression matrix}

\item{reactive_triset}{reactive value, a matrix of nx3 use for triselector}
}
\description{
Utility - extended figure control shiny module
}
\examples{
#' # library(shiny)
# library(shinyjs)
# library(Biobase)
# library(shinyWidgets)
# source("Git/R/module_triselector.R")
# dat <- readRDS("Dat/exampleEset.RDS")
# pd <- pData(dat)
# fd <- fData(dat)
# expr <- exprs(dat)
# ts <- trisetter(meta = pd, expr = expr, combine = "pheno")
# # ts <- stringr::str_split_fixed(colnames(pd), pattern = "\\\\|", n = 3)

# ui <- fluidPage(
#   attr4selector_ui("a4test")
# )

# server <- function(input, output, session) {
#   # observe(
#   #   print(k())
#   # )
# }#' 
# shinyApp(ui, server)

}
