% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_figureAttr4.R
\name{attr4selector_ui}
\alias{attr4selector_ui}
\title{Utility - extended figure control shiny ui}
\usage{
attr4selector_ui(id, circle = TRUE)
}
\arguments{
\item{id}{id}

\item{circle}{circle icon for dropdown manu}
}
\description{
Utility - extended figure control shiny ui
}
