% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_sparse_converter.R
\name{csc2list}
\alias{csc2list}
\title{convert a column compressed sparse matrix to a list}
\usage{
csc2list(x)
}
\arguments{
\item{x}{a matrix or CsparseMatrix object}
}
\value{
a sparse frame in data.frame
}
\description{
convert a column compressed sparse matrix to a list
}
