% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataTable.R
\name{dataTable_module}
\alias{dataTable_module}
\title{utility - dataTable shiny module}
\usage{
dataTable_module(input, output, session, reactive_data, selector = TRUE,
  columns = NULL, reactiveSelectorMeta = reactive(NULL),
  reactiveSelectorHeatmap = reactive(NULL), subset = c("row", "col",
  "none")[1])
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_data}{the data to be shown, a tabular objet}

\item{selector}{whether a selector should be added to the output}

\item{columns}{columns to show}

\item{reactiveSelectorMeta}{object returned by pdata or fdata}

\item{reactiveSelectorHeatmap}{object returned by heatmap}

\item{subset}{subset row or column for heatmap}
}
\description{
utility - dataTable shiny module
}
\examples{
# library(shiny)
# source("Git/R/module_triselector.R")
# 
# dat <- readRDS("Dat/exampleEset.RDS")
# expr <- exprs(dat)
# pdata <- pData(dat)
# fdata <- fData(dat)
# 
# ui <- fluidPage(
#   dataTable_ui("dttest")
# )
# 
# server <- function(input, output, session) {
#   callModule(dataTable_module, id = "dttest",  reactive_data = reactive(pdata))
# }
# 
# shinyApp(ui, server)
# 
# 
# ####
# ui <- fluidPage(
#   dataTable_ui("dttest")
# )
# 
# server <- function(input, output, session) {
#   callModule(dataTable_module, id = "dttest",  reactive_data = reactive(fdata))
# }
# 
# shinyApp(ui, server)
# 
# ###
# ui <- fluidPage(
#   dataTable_ui("dttest", selector = FALSE)
# )
# 
# server <- function(input, output, session) {
#   callModule(dataTable_module, id = "dttest",  reactive_data = reactive(expr), selector = FALSE)
# }
# 
# shinyApp(ui, server)

}
