% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataTable.R
\name{dataTable_ui}
\alias{dataTable_ui}
\title{utility - dataTable shiny UI}
\usage{
dataTable_ui(id, selector = TRUE)
}
\arguments{
\item{id}{id}

\item{selector}{where to show the selector}
}
\description{
utility - dataTable shiny UI
}
