% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataTableDownload.R
\name{dataTableDownload_module}
\alias{dataTableDownload_module}
\title{utility - dataTable for download shiny module}
\usage{
dataTableDownload_module(input, output, session, reactive_table,
  reactive_cols = reactive(NULL), prefix = "", pageLength = 10)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_table}{table to show}

\item{reactive_cols}{columns to be shown}

\item{prefix}{file name prefix}

\item{pageLength}{how many row per page}
}
\description{
A subset of columns can be shown, specified by reactive_cols. 
  The entire table will be downloaded.
}
\examples{
# source("R/module_triselector.R")
# library(shiny)
# library(stringr)
# 
# dat <- readRDS("../Dat/exampleEset.RDS")
# pdata <- pData(dat)
# 
# ui <- fluidPage(
#   dataTableDownload_ui("dtd")
# )
# server <- function(input, output, session) {
# }
# shinyApp(ui, server)

}
