% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataTableDownload.R
\name{dataTableDownload_ui}
\alias{dataTableDownload_ui}
\title{utility - dataTable for download shiny UI}
\usage{
dataTableDownload_ui(id, showTable = TRUE)
}
\arguments{
\item{id}{id}

\item{showTable}{logical, if the table should be shown}
}
\description{
utility - dataTable for download shiny UI
}
