% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_quickViews.R
\name{detect_quick_views}
\alias{detect_quick_views}
\title{Detect standard quick views from metadata column names}
\usage{
detect_quick_views(meta, triset = NULL)
}
\arguments{
\item{meta}{Feature or sample metadata.}

\item{triset}{Matrix returned by \code{\link{trisetter}}.}
}
\value{
A data.frame of quick-view definitions.
}
\description{
Detect standard quick views from metadata column names
}
\keyword{internal}
