% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{.e2EC50}
\alias{.e2EC50}
\title{convert e (inflection point) to EC50}
\usage{
.e2EC50(b, d, e, f)
}
\description{
convert e (inflection point) to EC50
}
\note{
Only has an effect when using LL.5 and LL2.5 model
}
