% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{.modelFormula}
\alias{.modelFormula}
\title{model fitted by drc}
\usage{
.modelFormula(x, b, c = 0, d = 1, e, f = 1)
}
\arguments{
\item{b}{Hill's slope. The Hill's slope refers to the steepness of the curve. It could either be positive or negative.}

\item{c}{Maximum asymptote. In an bioassay where you have a standard curve, this can be thought of as the response value for infinite standard concentration.}

\item{d}{Minimum asymptote. In a bioassay where you have a standard curve, this can be thought of as the response value at 0 standard concentration.}

\item{e}{Inflection point. The inflection point is defined as the point on the curve where the curvature changes direction or signs. e is the concentration of analyte where y=(c-d)/2.}

\item{f}{Asymmetry factor. When f=1 we have a symmetrical curve around inflection point and so we have a four-parameters logistic equation.}
}
\description{
model fitted by drc
}
\details{
func(x) = c + (d - c) / (1 + (x/e)^b)^f
}
