% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{getUPRefProteomeID}
\alias{getUPRefProteomeID}
\alias{downloadUPRefProteome}
\title{get uniprot reference proteome IDs}
\usage{
getUPRefProteomeID(
  domain = c("Eukaryota", "Archaea", "Bacteria", "Viruses")[1]
)

downloadUPRefProteome(
  id,
  domain = c("Eukaryota", "Archaea", "Bacteria", "Viruses")[1],
  destdir = "./"
)
}
\arguments{
\item{domain}{the domain, one of "Eukaryota", "Archaea", "Bacteria" or "Viruses"}

\item{id}{the UP id to download}

\item{destdir}{destination directory}
}
\value{
a character vector of UP ids

a character vector of UP ids
}
\description{
get uniprot reference proteome IDs

get uniprot reference proteome IDs
}
\section{Functions}{
\itemize{
\item \code{getUPRefProteomeID()}: get uniprot reference protein IDs

}}
