% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_ROC_PR.R
\name{draw_roc_pr}
\alias{draw_roc_pr}
\title{Drawing ROC and PR curve}
\usage{
draw_roc_pr(value, label)
}
\arguments{
\item{value}{a numerical vector indicates the predictions}

\item{label}{true class labels, could be two or more unique values}
}
\description{
Drawing ROC and PR curve
}
\examples{
v <- sort(rnorm(100))
l <- sample(1:2, size = 100, replace = TRUE)
draw_roc_pr(v, l)
l <- rep(c("b", "c", "a", "d"), each = 25)
draw_roc_pr(v, l)
draw_roc_pr(v, sample(l))
}
