% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ora.R
\name{enrichment_analysis_ui}
\alias{enrichment_analysis_ui}
\title{Utility enrichment analysis shiny ui}
\usage{
enrichment_analysis_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
Utility enrichment analysis shiny ui
}
