% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_fgsea.R
\name{enrichment_fgsea_module}
\alias{enrichment_fgsea_module}
\title{Utility - fgsea shiny module}
\usage{
enrichment_fgsea_module(input, output, session, reactive_featureData)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_featureData}{reactive feature data}
}
\description{
Utility - fgsea shiny module
}
\examples{
# library(shiny)
# source("Git/R/module_triselector.R")
# source("Git/R/auxi_fgsea.R")
# source("Git/R/module_barplotGsea.R")
# 

}
