% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_export_runExpressionSetViewer.R
\name{ExpressionSetViewer}
\alias{ExpressionSetViewer}
\title{Start ExpressionSetViewer}
\usage{
ExpressionSetViewer(dir, additionalTabs = NULL, esetLoader = readRDS,
  exprsGetter = exprs, pDataGetter = pData, fDataGetter = fData)
}
\arguments{
\item{dir}{directory to the ExpressionSet object. Only give the directoy
in this argument, not the .rds file.}

\item{additionalTabs}{additional tabs added to "Analyst" panel}

\item{esetLoader}{function to load the eset object, if an RDS file, should be "readRDS"}

\item{exprsGetter}{function to get the expression matrix from eset}

\item{pDataGetter}{function to get the phenotype data from eset}

\item{fDataGetter}{function to get the feature data from eset}
}
\description{
Start ExpressionSetViewer
}
\examples{
1
## To start the shiny app: 
# ExpressionSetViewer(
# 	ExpressionSetViewer(system.file("extdata", package = "ExpressionSetViewer"))
# )

}
