% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_stats.R
\name{exprspca}
\alias{exprspca}
\title{perform PCA and prepare for ExpressionSetViewer}
\usage{
exprspca(x, n = min(8, ncol(x) - 1), prefix = "PCA|All",
  fillNA = FALSE)
}
\arguments{
\item{x}{expression matrix, where rows are feature and samples are columns. Don't need to 
transpose the matrix.}

\item{n}{number of components to keep}

\item{prefix}{prefix of the names}

\item{fillNA}{Whether NA should be filled. If TRUE, the missing value will
be replaced by a constant a slightly lower than the min detected value (-log10(2)).
If set to false, na.omit will be called before pca}
}
\value{
a \code{data.frame} storing the PCA results
}
\description{
perform PCA and prepare for ExpressionSetViewer
}
\examples{
# reading expression
packdir <- system.file("extdata", package = "ExpressionSetViewer")
expr <- read.delim(file.path(packdir, "expressionMatrix.tsv"), stringsAsFactors = FALSE)
# call PCA
pc <- exprspca(expr)
head(pc$samples)
head(pc$features)
}
