% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{extendMetaData}
\alias{extendMetaData}
\alias{extendMetaData,ExpressionSet,data.frame-method}
\title{Add extra columns to the phenoData or featureData in ExpressionSet}
\usage{
extendMetaData(object, newData, where)

\S4method{extendMetaData}{ExpressionSet,data.frame}(object, newData, where = c("pData", "fData")[1])
}
\arguments{
\item{object}{an object of \code{ExpressionSet-class}}

\item{newData}{a \code{data.frame} containing the data to be added}

\item{where}{either "pData" or "fData"}
}
\value{
an object of \code{ExpressionSet-class}
}
\description{
Add extra columns to the phenoData or featureData in ExpressionSet
}
\note{
The attributes in the pheno data and feature data will be preserved
}
\examples{
est <- Biobase::ExpressionSet(assayData=matrix(runif(1000), nrow=100, ncol=10))
Biobase::pData(est)
est <- extendMetaData(est, data.frame(letter = letters[1:10]), where = "pData")
Biobase::pData(est)
}
