% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{extractParamDC}
\alias{extractParamDC}
\title{Extracting parameters from drc models}
\usage{
extractParamDC(mod, prefix = "ResponseCurve")
}
\arguments{
\item{mod}{a drc object}

\item{prefix}{for column header, the column will be named as prefix|curveid|curveparameter}
}
\description{
Extracting parameters from drc models
}
\note{
when LL2.X is used, e is estimated as log(e), this function will return e in linear scale instead.
}
