% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{extractParamDCList}
\alias{extractParamDCList}
\title{Extracting parameter from a list of drc object}
\usage{
extractParamDCList(x, prefix = "ResponseCurve")
}
\arguments{
\item{x}{a list of drc object}

\item{prefix}{for column header}
}
\value{
a data.frame
}
\description{
Extracting parameter from a list of drc object and return a data.frame,
which can be incorporated into the object visualized by omicsViewer
}
