% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_contTableStats.R
\name{factorIndependency}
\alias{factorIndependency}
\title{Independency test for two categorical vectors}
\usage{
factorIndependency(x, y)
}
\arguments{
\item{x}{a vector of character values}

\item{y}{the second vector of character values, the same length as x}
}
\description{
Independency test for two categorical vectors
}
\examples{
1
# # examples
# t1 <- sample(c('A', "B"), size = 30, replace = TRUE)
# t2 <- sample(c('a', "b"), size = 30, replace = TRUE)
# factorIndependency(t1, t2)
# t3 <- sample(c('a', "b", "c", 'd'), size = 30, replace = TRUE)
# factorIndependency(t1, t3)
}
