% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_contTableStats.R
\name{factorIndependency_module}
\alias{factorIndependency_module}
\title{utility - factorIndependency shiny module}
\usage{
factorIndependency_module(input, output, session, x, y,
  reactive_checkpoint = reactive(TRUE))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{x}{x, char vector}

\item{y}{y, char vector, same length as x}

\item{reactive_checkpoint}{checkpoint}
}
\description{
utility - factorIndependency shiny module
}
\examples{
1 #
# library(shiny)
# 
# ui <- fluidPage(
#   factorIndependency_ui("cti")
# )
# 
# server <- function(input, output, session) {
#   x0 <- reactive( sample(c('A', "B"), size = 30, replace = TRUE) )
#   y0 <- reactive( sample(c('a', "b", "c"), size = 30, replace = TRUE) )
#   callModule(factorIndependency_module, id = "cti", x = x0, y = y0)
# }
# 
# shinyApp(ui, server)
# 
# 
# t1 <- sample(c('A', "B"), size = 30, replace = TRUE)
# t2 <- sample(c('a', "b", "c"), size = 30, replace = TRUE)
# u <- factorIndependency(t1, t2)
# u <- u$cont.table
}
