% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_contTableStats.R
\name{factorIndependency_ui}
\alias{factorIndependency_ui}
\title{utility - factorIndependency shiny UI}
\usage{
factorIndependency_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
utility - factorIndependency shiny UI
}
