% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_feature_general.R
\name{feature_general_module}
\alias{feature_general_module}
\title{utility - feature general ui}
\usage{
feature_general_module(
  input,
  output,
  session,
  reactive_expr,
  reactive_i = reactive(NULL),
  reactive_highlight = reactive(NULL),
  reactive_phenoData,
  reactive_featureData
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_expr}{reactive expression matrix}

\item{reactive_i}{reactive row index to be highlighted}

\item{reactive_highlight}{reactive col index to be highlighted}

\item{reactive_phenoData}{reactive phenotype data}

\item{reactive_featureData}{reactive feature data}
}
\description{
utility - feature general ui
}
\examples{
#' # library(shiny)
# library(shinyBS)
# library(Biobase)
# dat <- readRDS("Dat/exampleEset.RDS")
# source("Git/R/module_triselector.R")
# source("Git/R/module_scatter.R")
# source("Git/R/module_contTableStats.R")
# source("Git/R/module_survival.R")
# source("Git/R/module_boxplot.R")
# source("Git/R/module_figureAttr4.R")
# source("Git/R/auxi_figureAttr4.R")
# 
# 
# ui <- fluidPage(
#   feature_general_ui("tres")
# )
# 
# server <- function(input, output, session) {
#   v <- callModule(feature_general_module, id = "tres",
#                   reactive_expr = reactive(exprs(dat)),
#                   # reactive_i = reactive(c(5, 6, 7)),
#                   reactive_i = reactive(c( 7)),
#                   reactive_highlight = reactive(c(3, 5, 10)),
#                   reactive_phenoData = reactive(pData(dat)),
#                   reactive_featureData = reactive(fData(dat)))
# }
# 
# shinyApp(ui, server)
}
