% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_fgsea.R
\name{fgsea0}
\alias{fgsea0}
\title{fgsea wrapper using binary matrix input}
\usage{
fgsea0(gs, stats, nperm = 100, gs_desc = NULL, feature_desc = NULL,
  ...)
}
\arguments{
\item{gs}{a matrix where rows are features and columns are gene sets. The zeros in the matrix 
indicate a gene is not a memeber of the gene sets in the column.}

\item{stats}{other parameter passed to fgsea}

\item{...}{other parameters passed to fgsea}
}
\description{
fgsea wrapper using binary matrix input
}
