% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_stats.R
\name{fillNA}
\alias{fillNA}
\title{Filling NA using a constant or half lowest detected value (for each protein) method}
\usage{
fillNA(
  x,
  maxfill = quantile(x, probs = 0.15, na.rm = TRUE),
  fillingFun = function(x) x - log10(2)
)
}
\arguments{
\item{x}{matrix with NA values}

\item{maxfill}{the maximum of filled value}

\item{fillingFun}{function to calculate half value
e.g.
function(x) x - log10(2) # default, when x is in log10 scale
function(x) x - 1 # half lowest detected value when x is in log2 scale
function(x) 0 # replace NA by 0}
}
\value{
a matrix without NAs
}
\description{
Filling NA using a constant or half lowest detected value (for each protein) method
}
\examples{
m <- matrix(rnorm(200), 20, 10)
m[sample(1:200, size = 20)] <- NA
mf <- fillNA(m)
}
