% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_quickViews.R
\name{get_quick_views}
\alias{get_quick_views}
\title{Read custom quick-view metadata}
\usage{
get_quick_views(meta)
}
\arguments{
\item{meta}{Feature or sample metadata.}
}
\value{
User-defined shortcut metadata, or NULL.
}
\description{
Read custom quick-view metadata
}
\keyword{internal}
