% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{getMQParams}
\alias{getMQParams}
\title{Parse mqpar.xml file}
\usage{
getMQParams(x)
}
\arguments{
\item{x}{the path to mqpar.xml file}
}
\description{
Getting the experimental informatione (TMT or label free) from mqpar.xml file.
}
