% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_databaseFormat.R
\name{getPData}
\alias{getPData}
\alias{getPData,SQLiteConnection-method}
\alias{getPData,ExpressionSet-method}
\alias{getPData,ANY-method}
\title{Extract the phenotype (sample) annotation of an omicsViewer dataset}
\usage{
getPData(x)

\S4method{getPData}{SQLiteConnection}(x)

\S4method{getPData}{ExpressionSet}(x)

\S4method{getPData}{ANY}(x)
}
\arguments{
\item{x}{an SQLite connection, an \code{ExpressionSet}, or any class
with its own \code{getPData} method (dispatched automatically)}
}
\value{
the sample annotation data.frame
}
\description{
Extract the phenotype (sample) annotation of an omicsViewer dataset
}
