% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_queryStringdb.R
\name{getStringId}
\alias{getStringId}
\title{Mapping ids to string ids}
\usage{
getStringId(genes, taxid = 9606, caller = "ExpressionSetViewer")
}
\arguments{
\item{genes}{a character vector of gene ids}

\item{taxid}{taxonomy ids}

\item{caller}{your identifier for string-db.org}
}
\description{
the string ids can be used as background in the string enrichment analysis
}
\note{
https://string-db.org/cgi/help.pl?subpage=api%23mapping-identifiers
}
\examples{
1
# gg = c('P04637', 'P00533', 'P04626', "Q8IYB3", "O75494", "Q9Y696")
# getStringId(gg)
}
