% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_export_prepEsetViewer.R
\name{gsAnnotIdList}
\alias{gsAnnotIdList}
\title{Gene set anntoation function}
\usage{
gsAnnotIdList(idList, gsIdMap, minSize = 5, maxSize = 500)
}
\arguments{
\item{idList}{list of protein IDs, e.g. list(c("ID1", "ID2"), c("ID13"), c("ID4", "ID8", "ID10"))}

\item{gsIdMap}{a data frame for geneset to id map, it has two columns
- id: the ID column
- term: annotation terms
e.g. 
gsIdMap <- data.frame(
  id = c("ID1", "ID2", "ID1", "ID2", "ID8", "ID10"),
  term = c("T1", ""T1", "T2", "T2", "T2", "T2"),
  stringsAsFactors = FALSE
  )}

\item{minSize}{minimum size of gene sets}

\item{maxSize}{maximum size of gene sets}
}
\description{
Gene set anntoation function
}
