% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_databaseFormat.R
\name{hasAttr}
\alias{hasAttr}
\title{Check whether an object has an attribute}
\usage{
hasAttr(x, attr.name)
}
\arguments{
\item{x}{the object}

\item{attr.name}{a character vector containing the name of attributes to be checked}
}
\value{
a logical value/vector has the same length as attr.name
}
\description{
Check whether an object has an attribute
}
