% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_databaseFormat.R
\name{hclust2str}
\alias{hclust2str}
\alias{str2hclust}
\title{Convert hclust object to/from single character}
\usage{
hclust2str(x)

str2hclust(x)
}
\arguments{
\item{x}{a character of length one or an hclust object}
}
\value{
a character stores the hclust object

a hclust object
}
\description{
Convert hclust object to/from single character
}
\note{
The $call element in hclust will not retained in the conversion. The conversion
  decrease the precision in $height element.
}
\examples{
# not for end users
# m <- matrix(rnorm(50), 25)
# hc <- hclust(dist(m))
# plot(hc)
# te <- hclust2str(hc)
# hc2 <- str2hclust(te)
# plot(hc2)

}
