% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_vectORA.R
\name{jaccardList}
\alias{jaccardList}
\title{Calculate Jaccard distance from a list}
\usage{
jaccardList(x)
}
\arguments{
\item{x}{a list}
}
\value{
an dist object
}
\description{
Calculate Jaccard distance from a list
}
