% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_module_data_space.R
\name{L1_data_space_module}
\alias{L1_data_space_module}
\title{Application level 1 module - data space}
\usage{
L1_data_space_module(
  id,
  expr,
  pdata,
  fdata,
  reactive_x_s = reactive(NULL),
  reactive_y_s = reactive(NULL),
  reactive_x_f = reactive(NULL),
  reactive_y_f = reactive(NULL),
  cormat = reactive(NULL),
  status = reactive(NULL),
  store = NULL
)
}
\arguments{
\item{id}{module id}

\item{expr}{reactive value; expression matrix}

\item{pdata}{reactive value; phenotype data}

\item{fdata}{reactive value; feature data}

\item{reactive_x_s}{pre-selected x axis for sample space}

\item{reactive_y_s}{pre-selected y axis for sample space}

\item{reactive_x_f}{pre-selected x axis for feature space}

\item{reactive_y_f}{pre-selected y axis for feature space}

\item{cormat}{reactive value; correlation matrix. if not given, calculated on the fly.}

\item{status}{intial status}
}
\description{
Application level 1 module - data space
}
