% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_shiny_misc.R
\name{nColors}
\alias{nColors}
\title{Generating k distinct colors}
\usage{
nColors(k, stop = FALSE)
}
\arguments{
\item{k}{a number between 2 to 60 tells how many distinct colors to use}

\item{stop}{logical; whether the function should return an error message if
k is not in the range of 2 to 60. Default FALSE, the function will return NULL.}
}
\value{
a vector of hex code for k colors or NULL
}
\description{
Mainly used in the shiny app to generate reproducible k distinct colors.
}
\examples{
nColors(5)
nColors(1, stop = FALSE)
}
