% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{normalize.softQuantile}
\alias{normalize.softQuantile}
\title{normalize soft quantile}
\usage{
normalize.softQuantile(x, probs = 0.5, sharedProtein = FALSE, ref = 1)
}
\arguments{
\item{x}{the input matrix, the columns of which were centered on quantiles, log10 transformed intensity}

\item{probs}{the quantiles used to controls/align. If probs is a length 1 numerical 
vector, the quantiles will aligned. If probs' length is > 1, the quantiles are calculated
and linear models are used to estimated slope and intersect between quantiles.}

\item{sharedProtein}{normalization based on the shared proteins between reference 
experiment each individual experiment}

\item{ref}{the columns name or index to specify the reference experiment, only used 
when sharedProtein = TRUE}
}
\description{
normalize soft quantile
}
