% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{normalize.totsum}
\alias{normalize.totsum}
\title{Normalize total sum}
\usage{
normalize.totsum(x)
}
\arguments{
\item{x}{a log10 transformed expression matrix}
}
\value{
a normalized matrix
}
\description{
Normalize total sum
}
\examples{
e1 <- matrix(rnorm(5000), 500, 10)
e1[, 6:10] <- e1[, 6:10]+3
boxplot(e1)
e2 <- normalize.totsum(x = e1)
boxplot(e2)
}
