% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{normalizeData}
\alias{normalizeData}
\title{normalize data row-wise or column-wise}
\usage{
normalizeData(
  x,
  colWise = c("None", "Median centering", "Median centering (shared ID)", "Total sum",
    "median centering + variance stablization")[1],
  rowWise = c("None", "Reference", "Batch mean", "Batch reference")[1],
  reference = NULL,
  batch = NULL
)
}
\arguments{
\item{x}{expression matrix where rows are features and columns are samples}

\item{colWise}{column wise normalization}

\item{rowWise}{row wise normalization method}

\item{reference}{index of reference samples}

\item{batch}{batch factor}
}
\description{
normalize data row-wise or column-wise
}
