% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{parseDatTerm}
\alias{parseDatTerm}
\title{Extract function annotation from uniprot .dat file}
\usage{
parseDatTerm(file, outputDir = NULL, ...)
}
\arguments{
\item{file}{the .dat or .dat.gz file}

\item{outputDir}{dir of output file}

\item{...}{other parameters passed to readLines}
}
\description{
Extract function annotation from uniprot .dat file
}
