% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{plotDC}
\alias{plotDC}
\title{Draw dose-response curves}
\usage{
plotDC(mod, ylab = "Abundance", lty = 2, pch = 19, cex = 1, ...)
}
\arguments{
\item{mod}{an \code{drc} object}

\item{ylab}{ylab in plot function}

\item{lty}{lty in plot function}

\item{pch}{pch in plot function}

\item{cex}{cex in plot function}

\item{...}{other arguments passed to plot function}
}
\description{
Draw dose-response curves
}
