% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{plotDCMat}
\alias{plotDCMat}
\title{Draw dose response curve given parameters in the omicsViewer object}
\usage{
plotDCMat(expr, pd, fd, featid, dose.var, curve.var, only.par = FALSE)
}
\arguments{
\item{expr}{expression matrix}

\item{pd}{phenotype data or colData}

\item{fd}{feature data or rowData}

\item{featid}{feature id to be visualized}

\item{dose.var}{the column header indicating the dose/time/concentration}

\item{curve.var}{the column header indicating the curve ids}

\item{return.par}{logical value. If true, no plot generated,
the function only returns the parameters of models.}
}
\description{
Draw dose response curve given the feature Data/rowData, phenotype data/colData 
and expression matrix. The function is usually used in shinyApp.
}
