% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_barplotGsea.R
\name{plotly_barplot}
\alias{plotly_barplot}
\title{plotly barplot}
\usage{
plotly_barplot(
  x,
  names,
  highlight = NULL,
  highlight_color = "red",
  highlight_width = 1,
  highlight_legend = "highlighted",
  background = NULL,
  background_color = "gray",
  background_width = 1,
  background_legend = "background",
  ylab = "ylab",
  xlab = "xlab",
  sort = c("none", "increasing", "decreasing")[1],
  tooltips = NULL,
  source = "plotlybarchart"
)
}
\arguments{
\item{x}{a numeric vector}

\item{names}{the names of the vector}

\item{highlight}{index for which value to be highlighted}

\item{highlight_color}{highlihgt color}

\item{highlight_width}{with highlight}

\item{highlight_legend}{legend for highlighted values, default "highlighted"}

\item{background}{index for background values}

\item{background_color}{background color}

\item{background_width}{background width}

\item{background_legend}{background legend}

\item{ylab}{label of y-axis}

\item{xlab}{label of x-axis}

\item{sort}{how the x should be sorted}

\item{tooltips}{tooltips for values}

\item{source}{passed to plotly, source name, use for capture mouse event}
}
\description{
plotly barplot
}
\examples{
1
# barplot utility - not for users
# # ## examples
# set.seed(100)
# n <- 100
# s_x <- rnorm(n)
# s_n <- paste(sample(letters, replace = TRUE, size = n), 1:n)
# plotly_barplot(x = s_x, names = s_n)
# plotly_barplot(x = s_x, names = s_n, sort = "inc")
# plotly_barplot(x = s_x, names = s_n, sort = "dec")
# 
# plotly_barplot(x = s_x, names = s_n, highlight = c(5, 6, 9, 10, NA), background = c(15, 18, 30))
# # 
# dat <- readRDS("Dat/exampleEset.RDS")
# fd <- fData(dat)
# plotly_barplot(fd$`t-test|OV_BR|md`, rownames(fd))
# plotly_barplot(fd$`t-test|OV_BR|md`, rownames(fd), 
#                highlight = c(1, 5, 10), highlight_width = 6,
#                sort = "inc")
# 
# plotly_barplot(fd$`t-test|OV_BR|md`, rownames(fd), 
#                highlight = c(1, 5, 10), highlight_width = 6,
#                background = c(20, 50, 70), background_width = 6,
#                sort = "inc")
}
