% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_boxplot.R
\name{plotly_boxplot}
\alias{plotly_boxplot}
\title{plotly boxplot wrapper}
\usage{
plotly_boxplot(
  x,
  i = NULL,
  highlight = NULL,
  ylab = "ylab",
  extvar = NULL,
  ylab.extvar = "ylab.extvar"
)
}
\arguments{
\item{x}{a matrix object}

\item{i}{index of which row in x should be highlighted}

\item{highlight}{index of which column in x to be highlighted}

\item{ylab}{y-axis label}

\item{extvar}{external variable, put on top of the boxplot}

\item{ylab.extvar}{y-axis label of external variable}
}
\description{
plotly boxplot wrapper
}
\examples{
1 # not for uesrs
# ## examples
# library(plotly)
# x <- cbind(matrix(rnorm(10000, mean = 3), 1000, 10), matrix(rnorm(20000), 1000, 20))
# x[sample(1:length(x), size = 0.3*length(x))] <- NA
# rownames(x) <- paste("R", 1:nrow(x), sep = "")
# colnames(x) <- paste("C", 1:ncol(x), sep = "")
# 
# plotly_boxplot(x)
# plotly_boxplot(x = x, i = 1)
# plotly_boxplot(x = x, i  = c(4, 20, 80))
# plotly_boxplot(x, i = 1:10)
# plotly_boxplot(x, i = 1:1000)
# 
# plotly_boxplot(x = x, highlight = c(1, 4, 5, 20))
# plotly_boxplot(x = x, i = 1, highlight = c(1, 4, 5, 20))
# plotly_boxplot(x = x, i  = c(4, 20, 80), highlight = c(1, 4, 5, 20))
# plotly_boxplot(x, i = 1:10, highlight = c(1, 4, 5, 20))
# plotly_boxplot(x, i = 1:1000, highlight = c(1, 4, 5, 20))
# 
# plotly_boxplot(x, i = c(4, 20, 80), extvar = 1:30)
}
