% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_quickViews.R
\name{prepare_quick_views}
\alias{prepare_quick_views}
\title{Combine custom and automatically detected quick views}
\usage{
prepare_quick_views(meta, triset = NULL)
}
\arguments{
\item{meta}{Feature or sample metadata.}

\item{triset}{Matrix returned by \code{\link{trisetter}}.}
}
\value{
A data.frame of available quick-view definitions.
}
\description{
Combine custom and automatically detected quick views
}
\keyword{internal}
