% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_export_prepEsetViewer.R
\name{prepEsetViewer}
\alias{prepEsetViewer}
\title{Prepare data that can be viewer by ExpressionSetViewer}
\usage{
prepEsetViewer(
  expr,
  pData,
  fData,
  PCA = TRUE,
  ncomp = min(8, ncol(expr)),
  pca.fillNA = TRUE,
  t.test = NULL,
  ttest.fillNA = FALSE,
  ...,
  gs = NULL,
  stringDB = NULL,
  surv = NULL
)
}
\arguments{
\item{expr}{expression matrix where the rows are feature and columns are samples, matrix should be log10 transformed}

\item{pData}{phenotype data}

\item{fData}{feature data}

\item{PCA}{pca}

\item{ncomp}{number of components to keep}

\item{pca.fillNA}{logical, whether the NA should be filled with a constant in PCA.}

\item{t.test}{How t-test should be done. it should be null (no t-test will be done)
or a nx3 matrix. The first column should be column headers in pData, the second and 
third columns should be two values in the columns of pData selected by the header 
(first column). The samples mapped to the two values are compared. If paired comparisons 
to be done, the orders of samples should be mapped.}

\item{ttest.fillNA}{logical, whether the NA should be filled with a constant in t-test.}

\item{...}{arguments passed to t.test, such as paired.}

\item{gs}{gene set data}

\item{stringDB}{string database}

\item{surv}{survival data}
}
\value{
an object of \code{ExpressionSet} that can be visualized using
\code{ExpressionSetViewer}
}
\description{
Prepare data that can be viewer by ExpressionSetViewer
}
\details{
expr should be a matrix with unique rownames and colnames. 
  pData and fData don't need to hav row names, row names will be assigned
  according to row names and columns of expr. 
  PCA and pairwise t-test could be done. The t.test argument should be a 
  nx3 matrix, the i
}
\examples{
packdir <- system.file("extdata", package = "ExpressionSetViewer")
# reading expression
expr <- read.delim(file.path(packdir, "expressionMatrix.tsv"), stringsAsFactors = FALSE)
colnames(expr) <- make.names(colnames(expr))
rownames(expr) <- make.names(rownames(expr))
# reading feature data
fd <- read.delim(file.path(packdir, "featureGeneral.tsv"), stringsAsFactors = FALSE)
# reading phenotype data
pd <- read.delim(file.path(packdir, "sampleGeneral.tsv"), stringsAsFactors = FALSE)

#  reading other datasets
drugData <- read.delim(file.path(packdir, "sampleDrug.tsv"))
# survival data
# this data is from cell line, the survival data are fake data to 
# show how to use the survival data in #' ExpressionSetViewer
surv <- read.delim(file.path(packdir, "sampleSurv.tsv"))
# gene set information
genesets <- read.delim(file.path(packdir, "featureGS.tsv"))

# Define t-test to be done
tests <- rbind(
c("Origin", "RE", "ME"),
c("Origin", "RE", "LE"),
c('TP53.Status', "MT", "WT")
)

# prepare column for stringDB query
strid <- sapply(strsplit(fd$Protein.ID, ";|-"), "[", 1)

###
d <- prepEsetViewer(
expr = expr, pData = pd, fData = fd, 
PCA = TRUE, pca.fillNA = TRUE,
t.test = tests, ttest.fillNA = FALSE, 
gs = genesets, stringDB = strid, surv = surv)

saveRDS(d, file = "dtest.RDS")
##  to open the viewer
# ExpressionSetViewer("./")
}
