% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_quickViews.R
\name{quick_badges_module}
\alias{quick_badges_module}
\title{Quick-view badge server logic}
\usage{
quick_badges_module(id, views, activeId)
}
\arguments{
\item{id}{Module id.}

\item{views}{Reactive data.frame of quick views.}

\item{activeId}{Reactive id of the currently active view.}
}
\value{
A reactive list containing the selected view and a monotonically
  increasing trigger counter.
}
\description{
Quick-view badge server logic
}
\keyword{internal}
