% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_quickViews.R
\name{quick_badges_ui}
\alias{quick_badges_ui}
\title{Quick-view badge UI}
\usage{
quick_badges_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A dynamic UI containing quick visualization buttons.
}
\description{
Quick-view badge UI
}
\keyword{internal}
