% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{read.proteinGroups.lf}
\alias{read.proteinGroups.lf}
\title{Read protein groups output of maxquant output and split
  it to columns}
\usage{
read.proteinGroups.lf(file)
}
\arguments{
\item{file}{Maxquant proteinGroup.txt file path}
}
\description{
Read protein groups output of maxquant output and split
  it to columns
}
