% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{read.proteinGroups}
\alias{read.proteinGroups}
\title{Reading proteinGroup table of MaxQuant output}
\usage{
read.proteinGroups(x, quant = c("LF", "TMT")[1])
}
\arguments{
\item{x}{the proteinGroup.txt file returned by MaxQuant search}

\item{quant}{the quantification method, LF or TMT}
}
\value{
a list of tables extracted from proteinGroups.txt file
}
\description{
A convenience function to read the proteinGroups table of MaxQuant 
  output. The function organize the result into different tables, e.g. iBAQ.
}
