% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{removeVarQC}
\alias{removeVarQC}
\title{Removing variance of reference samples}
\usage{
removeVarQC(x, ref, ...)
}
\arguments{
\item{x}{expression matrix}

\item{ref}{the index of reference samples}

\item{...}{if given, other parameters passed to normalize.softQuantile}
}
\description{
Removing variance of reference samples
}
