% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{rowshift}
\alias{rowshift}
\title{row-wise normalization}
\usage{
rowshift(x, batch, ref = NULL, useMean = FALSE)
}
\arguments{
\item{x}{a matrix where rows are genes/proteins and columns are samples, log10 transoformed intensity}

\item{batch}{a factor or vecter has the same length as ncol(x) to indicate
the batch assignment of samples}

\item{ref}{a logical vector has the same length as ncol(x) to indicated which
columns are the common references among batches. If it is NULL (by default), 
the mean of all channels will be used as batch reference.}

\item{useMean}{logical; whether to use means of batches, usually set to TRUE when no reference available}
}
\description{
row-wise normalization
}
