% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_export_runExpressionSetViewer.R
\name{runExpressionSetViewer}
\alias{runExpressionSetViewer}
\title{Start ExpressionSetViewer}
\usage{
runExpressionSetViewer(dir)
}
\arguments{
\item{dir}{directory to the ExpressionSet object}
}
\description{
Start ExpressionSetViewer
}
\examples{
1
# v <- testf("/media/share_baybioms/Projects/008_Bioinformatics/B032_ExpressionSetViewer/Dat/")

}
