% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sample_general.R
\name{sample_general_module}
\alias{sample_general_module}
\title{Utility sample general module}
\usage{
sample_general_module(
  input,
  output,
  session,
  reactive_phenoData,
  reactive_j = reactive(NULL)
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_phenoData}{reactive phenotype data}

\item{reactive_j}{index for which row in phenotype data should be highlighted/selected}
}
\description{
Utility sample general module
}
\examples{
#' # library(shiny)
# #
# source("Git/R/module_triselector.R")
# source("Git/R/module_scatter.R")
# source("Git/R/module_contTableStats.R")
# source("Git/R/module_survival.R")
# source("Git/R/module_figureAttr4.R")
# source("Git/R/auxi_figureAttr4.R")
# 
# dat <- readRDS("Dat/exampleEset.RDS")
# pd <- pData(dat)
# 
# ui <- fluidPage(
#   sample_general_ui("sample_general")
# )
# 
# server <- function(input, output, session) {
# }
# 
# shinyApp(ui, server)

}
