% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_queryStringdb.R
\name{stringGSA}
\alias{stringGSA}
\title{Performing gene set analysis using string-db}
\usage{
stringGSA(
  genes,
  taxid = 9606,
  background = NULL,
  backgroundStringId = FALSE,
  caller = "ExpressionSetViewer"
)
}
\arguments{
\item{genes}{a character vector of gene ids}

\item{taxid}{taxonomy ids (species NCBI identifier)}

\item{background}{the background genes}

\item{backgroundStringId}{logical value, whether the identifier given in background is 
already stringid, only used when background != NULL. You can use \code{\link{getStringId}}
to convet your id to string id.}

\item{caller}{your identifier for string-db.org}
}
\description{
Performing gene set analysis using string-db
}
\note{
https://string-db.org/cgi/help.pl?subpage=api%23getting-functional-enrichment
}
\examples{
1
# gg = c('P04637', 'P00533', 'P04626', "Q8IYB3", "O75494", "Q9Y696")
# u <- stringGSA(gg)
}
