% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_queryStringdb.R
\name{stringNetwork}
\alias{stringNetwork}
\title{Retrieve string network}
\usage{
stringNetwork(genes, taxid = 9606, caller = "ExpressionSetViewer")
}
\arguments{
\item{genes}{the gene ids}

\item{taxid}{taxonomy ids}

\item{caller}{your identifier for string-db.org}
}
\description{
get string network
}
