% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_survival.R
\name{survival_module}
\alias{survival_module}
\title{Utility survival KM module}
\usage{
survival_module(input, output, session, reactive_resp, reactive_strata,
  reactive_checkpoint = reactive(TRUE))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_resp}{reponse value, in the format 1, 345, 345+, 23, 45, 355+}

\item{reactive_strata}{strata variable}

\item{reactive_checkpoint}{checkpoint}
}
\description{
Utility survival KM module
}
\examples{
#' # library(shiny)
# library(survminer)
# library(survival)
# 
# ui <- fluidPage(
#   survival_ui("surv")
# )
# 
# server <- function(input, output, session) {
# }
# 
# # significant
# v <- runif(n = 1000, 1, 1000)
# t <- paste0(v, sample(c("", "+"), replace = TRUE, size = 1000))
# strata <- c("a", "b")[as.integer(v < 500)+1]
# shinyApp(ui, server)
# 
# # insignificant
# v <- runif(n = 1000, 1, 1000)
# t <- paste0(v, sample(c("", "+"), replace = TRUE, size = 1000))
# strata <- sample(c("a", "b"), replace = TRUE, size = 1000)
# shinyApp(ui, server)

}
