% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapshiny_auxi.R
\name{tooltipsShinyPlot}
\alias{tooltipsShinyPlot}
\alias{shinyPlotTooltips}
\title{Tooltips for shiny plot, Module}
\usage{
shinyPlotTooltips(input, output, session, points)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{points}{legend}
}
\description{
Tooltips for shiny plot, Module
}
\details{
modified from https://stackoverflow.com/questions/39346273/how-to-place-shiny-panel-near-the-pointer-mouse/50623683
}
\examples{
1
# require(shiny)
# require(ggplot2)
# ui <- shinyUI(fluidPage(
# shinyPlotTooltipsUI("stp"),
# plotOutput( "myplot", hover = hoverOpts(id ="myplot_hover") )
# ))

# server <- shinyServer(function(input, output, session) {
  
# output$myplot <- renderPlot({
# ggplot(mtcars) + geom_point(aes(mpg,cyl))
# })

## Create reactive variable
#points <- reactive({
# req(input$myplot_hover)
# unlist(nearPoints(mtcars, input$myplot_hover, maxpoints=1))
# })
  
# callModule(shinyPlotTooltips, id = "stp", points = points)
# })

# shinyApp(ui, server)

}
