% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_triselector.R
\name{triselector_ui}
\alias{triselector_ui}
\title{The three-step selector - the ui function}
\usage{
triselector_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
The three-step selector - the ui function
}
