% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_figureAttr4.R
\name{trisetter}
\alias{trisetter}
\title{Create a nx3 matrix that can be use for triselector given a meta and expression table}
\usage{
trisetter(meta, expr = NULL, combine)
}
\arguments{
\item{meta}{a meta data, usually either phenotype data or feature data}

\item{expr}{expression matrix, optional.}

\item{combine}{how the meta and expression to be combined. Should be either 
"pheno" or "feature' or "none".}
}
\value{
a nx3 matrix
}
\description{
only used inside reactive
}
