% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_funs.R
\name{validMQFolder}
\alias{validMQFolder}
\title{MQ folder validator
Validate whether a folder is a MQ output folder}
\usage{
validMQFolder(dir)
}
\arguments{
\item{dir}{the directory to check}
}
\value{
a list containing the info about MQ folder check
}
\description{
MQ folder validator
Validate whether a folder is a MQ output folder
}
\details{
from the root level, these files exist:
   mqpar.xml
   [[combined/]txt/]proteinGroups.txt
}
