% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapshiny_auxi.R
\name{value2color}
\alias{value2color}
\title{convert value to colors}
\usage{
value2color(x, n = 10)
}
\arguments{
\item{x}{a vector of numeric, character or factor}

\item{n}{number of distinct values, only used when x is a numeric vector}
}
\description{
convert value to colors
}
