% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_figureAttr4.R
\name{varSelector}
\alias{varSelector}
\title{variable selector}
\usage{
varSelector(x, expr, meta, alternative = NULL)
}
\arguments{
\item{x}{variable return by triselector, a list of length three named as "analysis",
"subset" and "variable'}

\item{expr}{the expression matrix}

\item{meta}{a meta matrix}

\item{alternative}{alternative value to be returned when nothing to select}
}
\value{
the selected values in input argument \code{x}
}
\description{
variable selector
}
