% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_vectORA.R
\name{vectORA.core}
\alias{vectORA.core}
\title{Internal function used by vectORA}
\usage{
vectORA.core(n.overlap, n.de, n.gs, n.bkg, unconditional.or = TRUE,
  mtc.method = "fdr")
}
\arguments{
\item{n.overlap}{the number of overlap between de and gs. The number of white balls drawn 
without replacement from an urn which contains both black and white balls (compared to hyper).}

\item{n.de}{the number of DE gene. The number of balls drawn from the urn (compared to hyper).}

\item{n.gs}{the size of gene set. The number of white balls in the urn (compared to hyper).}

\item{n.bkg}{the background size.}

\item{unconditional.or}{calculate odds ratio using Maximum Likelihood Estimate (the sample odds ratio). 
Note that the conditional Maximum Likelihood Estimate (MLE) is used in fisher.test.}

\item{mtc.method}{multiple test correction methods, passed to p.adjust function}
}
\description{
Internal function used by vectORA
}
\examples{
xq <- rbind(c(4, 2, 4),
            c(20, 40, 10),
            c(11, 234, 10),
            c(200, 1000, 100))

vectORA.core(xq[1, ], xq[2, ], xq[3, ], xq[4, ])
vectORA.core(xq[1, ], xq[2, ], xq[3, ], xq[4, ], unconditional.or = TRUE)

# fisher's test
t(apply(xq, 2, function(x1) {
  v <- fisher.test(rbind(c(x1[1], x1[2]-x1[1]), c(x1[3]-x1[1], x1[4] - x1[2] - x1[3] + x1[1])), alternative = "greater")
  c(p.value = v$p.value, v$estimate)
}))
}
