% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_vectORA.R
\name{vectORA}
\alias{vectORA}
\title{Vectorized over-representation analysis using fisher's exact test}
\usage{
vectORA(
  gs,
  i,
  background = NA,
  minOverlap = 3,
  minSize = 5,
  maxSize = Inf,
  gs_desc = NULL,
  feature_desc = NULL,
  unconditional.or = TRUE,
  mtc.method = "fdr",
  sort = c("none", "p.value", "OR")[1]
)
}
\arguments{
\item{gs}{a matrix where rows are features and columns are gene sets. The zeros in the matrix 
indicate a gene is not a memeber of the gene sets in the column.}

\item{i}{the row number selected from the gs}

\item{background}{An integer to indicate the size of background; if NA, background = nrow(gs)}

\item{minOverlap}{the minimum required overlap between gene set and gene list, if the overlap is lower
than this value, no test would be done on this gene set}

\item{minSize}{the minimum size of gene sets should be tested}

\item{maxSize}{the maximum size of genesets should be tested}

\item{gs_desc}{description of gene sets, a name character vector storing the description of
of the gene set, the names should be the same as colnames of gs}

\item{feature_desc}{feature description, e.g. gene symbols, names}

\item{unconditional.or}{logical. Whether calculate odds ratio using Maximum Likelihood Estimate (the sample odds ratio). 
Note that the conditional Maximum Likelihood Estimate (cMLE, set this parameter to FALSE) is used in fisher.test.}

\item{mtc.method}{multiple test correction methods, passed to p.adjust function}

\item{sort}{could be one of c("none", "p.value", "OR") to indicate how the result should be sorted.}
}
\description{
Vectorized over-representation analysis using fisher's exact test
}
\examples{
1
# library(Biobase)
# dat <- readRDS("Dat/exampleEset.RDS")
# fd <- fData(dat)
# fdgs <- fd[, grep("^GS\\|", colnames(fd))]
# ir <- which(fd$`t-test|OV_BR|pval` < 0.05)
# res <- vectORA(gs = fdgs, i = ir, unconditional.or = FALSE)
# res2 <- vectORA(gs = fdgs, i = ir, unconditional.or = TRUE)
}
