% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0_module_app.R
\name{app_module}
\alias{app_module}
\title{Application level 0 module}
\usage{
app_module(input, output, session, dir)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{dir}{reactive; directory containing the .RDS file of ExpressionSet}
}
\description{
Application level 0 module
}
