% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0_module_app.R
\name{app_module}
\alias{app_module}
\title{Application level 0 module}
\usage{
app_module(input, output, session, dir, additionalTabs = NULL,
  esetLoader = readRDS, exprsGetter = exprs, pDataGetter = pData,
  fDataGetter = fData)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{dir}{reactive; directory containing the .RDS file of ExpressionSet}

\item{additionalTabs}{additional tabs added to "Analyst" panel}

\item{esetLoader}{function to load the eset object, if an RDS file, should be "readRDS"}

\item{exprsGetter}{function to get the expression matrix from eset}

\item{pDataGetter}{function to get the phenotype data from eset}

\item{fDataGetter}{function to get the feature data from eset}
}
\description{
Application level 0 module
}
