% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0_module_app.R
\name{app_ui}
\alias{app_ui}
\title{application level 0 UI}
\usage{
app_ui(id, showDropList = TRUE, activeTab = "Feature")
}
\arguments{
\item{id}{id}

\item{showDropList}{logical; whether to show the dropdown list to select RDS file, if the 
ESVGlobalObj is given, this should be set to "FALSE"}

\item{activeTab}{one of "Feature", "Feature table", "Sample", "Sample table", "Heatmap"}
}
\description{
application level 0 UI
}
