% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0_module_app.R
\name{app_ui}
\alias{app_ui}
\title{Application level 0 UI}
\usage{
app_ui(id, showDropList = TRUE, activeTab = "Feature")
}
\arguments{
\item{id}{id}

\item{showDropList}{logical; whether to show the dropdown list to select RDS file, if the 
ESVObj is given, this should be set to "FALSE"}

\item{activeTab}{one of "Feature", "Feature table", "Sample", "Sample table", "Heatmap"}
}
\value{
a list of UI components
}
\description{
Function should only be used for the developers
}
\examples{
if (interactive()) {
  dir <- system.file("extdata", package = "ExpressionSetViewer")
  server <- function(input, output, session) {
    callModule(app_module, id = "app", dir = reactive(dir))
  }
  ui <- fluidPage(
    app_ui("app")
  )
  shinyApp(ui = ui, server = server)
}
}
