% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_databaseFormat.R
\name{asEsetWithAttr}
\alias{asEsetWithAttr}
\title{Convert SummarizedExperiment to ExpressionSet retaining all attributes}
\usage{
asEsetWithAttr(x)
}
\arguments{
\item{x}{an object of class SummarizedExperiment}
}
\value{
an object of class ExpressionSet
}
\description{
Convert SummarizedExperiment to ExpressionSet retaining all attributes
}
