% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_export_prepEsetViewer.R
\name{correlationAnalysis}
\alias{correlationAnalysis}
\title{Correlating a expression matrix with phenotypical variables}
\usage{
correlationAnalysis(expr, pheno, min.value = 12, prefix = "Cor")
}
\arguments{
\item{expr}{expression matrix, rows are the features (e.g. proteins), colnames are the samples}

\item{pheno}{a matrix stores the phenotypical variable to be correlated with expr. The rows 
should be samples, columns are phenotypical variables.}

\item{min.value}{the minimum number of samples required in the correlation analysis}

\item{prefix}{prefix used for columns headers}
}
\description{
Correlating a expression matrix with phenotypical variables
}
\examples{
e1 <- matrix(rnorm(500), 50, 10)
rownames(e1) <- paste0("FT", 1:50)
p1 <- matrix(rnorm(50), 10, 5)
colnames(p1) <- paste0("PH", 1:5)
colnames(e1) <- rownames(p1) <- paste0("S", 1:10)
correlationAnalysis(expr = e1, pheno = p1, min.value = 8)
}
