% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataTableDownload.R
\name{dataTableDownload_ui}
\alias{dataTableDownload_ui}
\title{utility - dataTable for download shiny UI}
\usage{
dataTableDownload_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
utility - dataTable for download shiny UI
}
