% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ora.R
\name{enrichment_analysis_module}
\alias{enrichment_analysis_module}
\title{Utility enrichment analysis shiny module}
\usage{
enrichment_analysis_module(
  input,
  output,
  session,
  reactive_featureData,
  reactive_i
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_featureData}{reactive feature data}

\item{reactive_i}{reactive index of rows to be selected (for ORA)}
}
\description{
Utility enrichment analysis shiny module
}
\examples{
#' # source("Git/R/auxi_fgsea.R")
# source("Git/R/auxi_vectORA.R")
# source("Git/R/module_barplotGsea.R")
}
