% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_export_runExpressionSetViewer.R
\name{ExpressionSetViewer}
\alias{ExpressionSetViewer}
\title{Start ExpressionSetViewer}
\usage{
ExpressionSetViewer(dir, additionalTabs = NULL, filePattern = ".RDS$",
  esetLoader = readRDS, exprsGetter = exprs, pDataGetter = pData,
  fDataGetter = fData, defaultAxisGetter = function(x, what = c("sx",
  "sy", "fx", "fy")[1]) attr(x, what), appName = "ExpressionSetViewer",
  appVersion = packageVersion("ExpressionSetViewer"))
}
\arguments{
\item{dir}{directory to the ExpressionSet object. Only give the directoy
in this argument, not the .rds file.}

\item{additionalTabs}{additional tabs added to "Analyst" panel}

\item{filePattern}{file pattern to be displayed.}

\item{esetLoader}{function to load the eset object, if an RDS file, should be "readRDS"}

\item{exprsGetter}{function to get the expression matrix from eset}

\item{pDataGetter}{function to get the phenotype data from eset}

\item{fDataGetter}{function to get the feature data from eset}

\item{defaultAxisGetter}{function to get the default axes to be visualized. It should be a function with two 
arguments: x - the object loaded to the viewer; what - one of "sx", "sy", "fx" and "fy", representing the 
sample space x-axis, sample space y-axis, feature space x-axis and feature space y-axis respectively.}

\item{appName}{name of the application}

\item{appVersion}{version of the application}
}
\description{
Start ExpressionSetViewer
}
\examples{
1
## To start the shiny app: 
# ExpressionSetViewer(
# 	system.file("extdata", package = "ExpressionSetViewer")
# )

}
