% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_export_runExpressionSetViewer.R
\name{ExpressionSetViewer}
\alias{ExpressionSetViewer}
\title{Start ExpressionSetViewer}
\usage{
ExpressionSetViewer(dir)
}
\arguments{
\item{dir}{directory to the ExpressionSet object. Only give the directoy
in this argument, not the .rds file.}
}
\description{
Start ExpressionSetViewer
}
\examples{
1
## To start the shiny app: 
# ExpressionSetViewer(
# 	ExpressionSetViewer(system.file("extdata", package = "ExpressionSetViewer"))
# )

}
