% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{extractParamDC}
\alias{extractParamDC}
\title{Extracting parameters from drc models}
\usage{
extractParamDC(mod, prefix = "ResponseCurve")
}
\arguments{
\item{mod}{a drc object}

\item{prefix}{for column header, the column will be named as prefix|curveid|curveparameter}
}
\description{
Extracting parameters from drc models
}
\details{
modelFormula <- function(x, b, c, d, e, f) {
  c + (d - c) / (1 + (x/e)^b)^f
}
d - Minimum asymptote. In a bioassay where you have a standard curve, this can be thought of as the response value at 0 standard concentration.
c - Maximum asymptote. In an bioassay where you have a standard curve, this can be thought of as the response value for infinite standard concentration.
e - Inflection point. The inflection point is defined as the point on the curve where the curvature changes direction or signs. e is the concentration of analyte where y=(c-d)/2.
b - Hill's slope. The Hill's slope refers to the steepness of the curve. It could either be positive or negative.
f - Asymmetry factor. When f=1 we have a symmetrical curve around inflection point and so we have a four-parameters logistic equation.
}
\note{
when LL2.X is used, e is estimated as log(e), this function will return e in linear scale instead.
}
