% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_fgsea.R
\name{fgsea0}
\alias{fgsea0}
\title{Wrapper of fgsea function to take binary gene set matrix as input}
\usage{
fgsea0(gs, stats, nperm = 100, gs_desc = NULL, ...)
}
\arguments{
\item{gs}{a matrix where rows are features and columns are gene sets. The zeros in the matrix 
indicate a gene is not a memeber of the gene sets in the column.}

\item{stats}{other parameter passed to fgsea}

\item{nperm}{number of permutation}

\item{gs_desc}{description of gene sets, it should be a named vector and the names
should be the same as colnames(gs)}

\item{...}{other parameters passed to fgsea}
}
\description{
Wrapper of fgsea function to take binary gene set matrix as input
}
\examples{
## not for users
# library(fgsea)
# library(Biobase)
# dat <- readRDS("Dat/exampleEset.RDS")
# fd <- fData(dat)
# fdgs <- fd[, grep("^GS\\\\|", colnames(fd))]
# res <- fgsea0(fdgs, stats = fd$`t-test|OV_BR|md`, nperm = 1000, minSize = 5, maxSize = 500)
# res <- fgsea0(fdgs, stats = fd$`t-test|OV_BR|md`, nperm = 1000, minSize = 5, maxSize = 500, gs_desc = colnames(fdgs))
}
