% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_databaseFormat.R
\name{getExprs}
\alias{getExprs}
\alias{getExprs,SQLiteConnection-method}
\alias{getExprs,ExpressionSet-method}
\alias{getExprs,xcmsFeatureSet-method}
\alias{getExprs,ANY-method}
\title{Extract the expression matrix of an omicsViewer dataset}
\usage{
getExprs(x)

\S4method{getExprs}{SQLiteConnection}(x)

\S4method{getExprs}{ExpressionSet}(x)

\S4method{getExprs}{xcmsFeatureSet}(x)

\S4method{getExprs}{ANY}(x)
}
\arguments{
\item{x}{an SQLite connection, an \code{ExpressionSet}, or any class
with its own \code{getExprs} method (dispatched automatically)}
}
\value{
the (features x samples) expression matrix
}
\description{
Extract the expression matrix of an omicsViewer dataset
}
