% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapshinyApp.R
\name{iheatmap}
\alias{iheatmap}
\alias{iheatmapInput}
\alias{iheatmapOutput}
\alias{iheatmapLegend}
\alias{iheatmapClear}
\alias{iheatmapModule}
\title{interactive heatmap}
\usage{
iheatmap(x, fData = NULL, pData = NULL, impute = FALSE)

iheatmapInput(id)

iheatmapOutput(id)

iheatmapLegend(id)

iheatmapClear(id)

iheatmapModule(
  input,
  output,
  session,
  mat,
  pd,
  fd,
  rowDendrogram = reactive(NULL)
)
}
\arguments{
\item{x}{a matrix object or ExpressionSet}

\item{fData}{feature data, ignored if x is an ExpressionSet}

\item{pData}{phenotype data, ignored if x is an ExpressionSet}

\item{impute}{whether impute the expression matrix}

\item{id}{id}

\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{mat}{expression matrix}

\item{pd}{phenotype data}

\item{fd}{feature data}

\item{rowDendrogram}{row dendrogram list}
}
\description{
interactive heatmap

iheatmap input

iheatmap Output

iheatmap legend

iheatmap clear function

iheatmap module
}
\section{Functions}{
\itemize{
\item \code{iheatmapInput}: 

\item \code{iheatmapOutput}: 

\item \code{iheatmapLegend}: 

\item \code{iheatmapClear}: iheatmap clear
}}

