% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_module_data_space.R
\name{L1_data_space_module}
\alias{L1_data_space_module}
\title{Application level 1 module - data space}
\usage{
L1_data_space_module(input, output, session, expr, pdata, fdata)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{expr}{reactive value; expression matrix}

\item{pdata}{reactive value; phenotype data}

\item{fdata}{reactive value; feature data}
}
\description{
Application level 1 module - data space
}
