% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_module_data_space.R
\name{L1_data_space_ui}
\alias{L1_data_space_ui}
\title{Application level 1 ui - data space}
\usage{
L1_data_space_ui(id, activeTab = "Feature")
}
\arguments{
\item{id}{id}

\item{activeTab}{one of "Feature", "Feature table", "Sample", "Sample table", "Heatmap"}
}
\description{
Application level 1 ui - data space
}
