% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_meta_scatter.R
\name{meta_scatter_module}
\alias{meta_scatter_module}
\title{Utility - scatter plot for meta shiny module}
\usage{
meta_scatter_module(
  id,
  reactive_meta = reactive(NULL),
  reactive_expr = reactive(NULL),
  combine = c("pheno", "feature"),
  source = "plotlyscattersource",
  reactive_x = reactive(NULL),
  reactive_y = reactive(NULL),
  reactive_status = reactive(NULL)
)
}
\arguments{
\item{id}{Character. Namespace ID for the Shiny module.}

\item{reactive_meta}{reactive meta data, phenotype data or feature data}

\item{reactive_expr}{reactive expression data}

\item{combine}{how to combine the expression and meta data, pheno or feature?}

\item{source}{source id for plotly object}

\item{reactive_x}{reactive value for pre-selected x-aixs}

\item{reactive_y}{reactive value for pre-selected y-aixs}

\item{reactive_status}{the status of scatter plot, e.g. x-, y-axis, color variable, shape variable, etc.}
}
\description{
Utility - scatter plot for meta shiny module
}
