% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_meta_scatter.R
\name{meta_scatter_ui}
\alias{meta_scatter_ui}
\title{Utility - scatter plot for meta shiny ui}
\usage{
meta_scatter_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
Utility - scatter plot for meta shiny ui
}
