% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_stats.R
\name{multi.t.test}
\alias{multi.t.test}
\title{Function to perform pairwse t-test}
\usage{
multi.t.test(x, pheno, compare = NULL, fillNA = FALSE, ...)
}
\arguments{
\item{x}{an expression matrix, log10 transformed}

\item{pheno}{phenotype data of x, nrow(pheno) must equal ncol(x)}

\item{compare}{comparison want to do. If not null, it a nx3 matrix. 
The first column should be column headers in pData, the second and 
third columns should be two values in the columns of pData selected by the header 
(first column). The samples mapped to the two values are compared. If paired comparisons 
to be done, the orders of samples should be mapped}

\item{fillNA}{Whether NA should be filled. If TRUE, the missing value will
be replaced by a constant a slightly lower than the min detected value (-log10(2))}

\item{...}{other parameters passed to t.test}
}
\value{
a \code{data.frame} stores the t-test results
}
\description{
Function to perform pairwse t-test
}
\examples{
# reading expression
packdir <- system.file("extdata", package = "ExpressionSetViewer")
expr <- read.delim(file.path(packdir, "expressionMatrix.tsv"), stringsAsFactors = FALSE)
# reading phenotype data
pd <- read.delim(file.path(packdir, "sampleGeneral.tsv"), stringsAsFactors = FALSE)

## Single t-test
head(pd)
# define comparisons
tests <- c("Origin", "RE", "ME")
tres <- multi.t.test(x = expr, pheno = pd, compare = tests)

## multiple t-test
head(pd)
# define comparisons
tests <- rbind(
c("Origin", "RE", "ME"),
c("Origin", "RE", "LE"),
c('TP53.Status', "MT", "WT")
)
tres <- multi.t.test(x = expr, pheno = pd, compare = tests)
}
