% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_doseCurve.R
\name{plotDC}
\alias{plotDC}
\title{Draw dose-response curves}
\usage{
plotDC(mod, ylab = "Abundance", lty = 2, pch = 19, cex = 1, logx = FALSE)
}
\arguments{
\item{mod}{an \code{drc} object}

\item{ylab}{ylab in plot function}

\item{lty}{lty in plot function}

\item{pch}{pch in plot function}

\item{cex}{cex in plot function}

\item{logx}{whether the x-axis should be in log scale}
}
\description{
Draw dose-response curves
}
