% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatter.R
\name{plotly_scatter}
\alias{plotly_scatter}
\title{Utility plotly scater}
\usage{
plotly_scatter(
  x,
  y,
  xlab = "",
  ylab = "ylab",
  color = "",
  shape = "",
  size = 10,
  tooltips = NULL,
  regressionLine = FALSE,
  source = "scatterplotlysource",
  sizeRange = c(5, 15),
  highlight = NULL,
  highlightName = "Highlighted",
  vline = NULL,
  hline = NULL,
  rect = NULL,
  drawButtonId = NULL
)
}
\arguments{
\item{x}{x values}

\item{y}{y values, same legnth as x}

\item{xlab}{label x-axis}

\item{ylab}{label y-axis}

\item{color}{color vector, same length as x and y}

\item{shape}{shape vector, same length as x and y}

\item{size}{size vector, same length as x and y}

\item{tooltips}{tooltip vector, same length as x and y}

\item{regressionLine}{show or not show regressionLine}

\item{source}{plotly source of plot}

\item{sizeRange}{size range of points}

\item{highlight}{whic value to highlight}

\item{highlightName}{legend for highlighted values}

\item{vline}{vertical line, not implemented}

\item{hline}{horizontal line, not implmented}

\item{rect}{rectangle coordinate in the form c(x0, x1, y0, y1)}

\item{drawButtonId}{not used currently}
}
\description{
Utility plotly scater
}
\examples{
#' # scatter plot
# x <- rnorm(30)
# y <- x + rnorm(30, sd = 0.5)
# plotly_scatter(x , y)
# plotly_scatter(x, y, tooltips = paste("A", 1:30))
# plotly_scatter(x, y, tooltips = paste("A", 1:30), regressionLine = TRUE, rect = c(x0 = -1, x1 = 1, y0 = -1, y1 = 1), vline = 1, hline = -1)
# plotly_scatter(x, y, tooltips = paste("A", 1:30), regressionLine = TRUE, color = rep(c("a", "b"), #' time = c(10, 20)))
# plotly_scatter(x, y, tooltips = paste("A", 1:30), regressionLine = TRUE,
#                )
# 
# ### beeswarm horizontal
# x <- rnorm(30)
# y <- sample(c("x", "y", "z"), size = 30, replace = TRUE)
# plotly_scatter(x = x , y = y)
# plotly_scatter(x, y, tooltips = paste("A", 1:30))
# plotly_scatter(x, y, tooltips = paste("A", 1:30), regressionLine = TRUE)
# plotly_scatter(x, y, tooltips = paste("A", 1:30), regressionLine = TRUE, 
# 

### beeswarm vertical
# 
# dt <- plotly_scatter(x = pd$`General|All|Origin`,
#                      y = pd$`General|All|Doubleing Time`,
#                      color = pd$`General|All|Origin`)
# 
# dt <- plotly_scatter(y = pd$`General|All|Origin`,
#                      x = pd$`General|All|Doubleing Time`,
#                      color = pd$`General|All|Origin`, highlight = c(5, 8, 10))
# dt
# # dt$fig
}
