% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatter.R
\name{plotly_scatter_ui}
\alias{plotly_scatter_ui}
\title{Utility plotly scatter ui}
\usage{
plotly_scatter_ui(id, height = "400px")
}
\arguments{
\item{id}{id}

\item{height}{figure height}
}
\description{
Utility plotly scatter ui
}
