% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_gsAnnotatedDF.R
\name{readESVObj}
\alias{readESVObj}
\title{Read the object of ExpressetSet to be visualized using ExpressionSetViewer}
\usage{
readESVObj(x)
}
\arguments{
\item{x}{a name of the file to be read, passed to \link{readRDS}}
}
\value{
An object of ExpressionSet
}
\description{
This function is similar to readRDS. The only difference is
this function will convert gene set data to data.frame format, if they are 
given in a matrix format.
}
\examples{
file <- system.file("extdata/demo.RDS", package = "ExpressionSetViewer")
obj <- readESVObj(file)
}
