% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_queryStringdb.R
\name{stringNetwork}
\alias{stringNetwork}
\title{Retrieve string network}
\usage{
stringNetwork(genes, taxid = 9606, caller = "ExpressionSetViewer")
}
\arguments{
\item{taxid}{taxonomy ids}

\item{caller}{your identifier for string-db.org}

\item{stringIds}{a character vector of gene ids}

\item{esetIds}{expression set IDs= NULL}
}
\description{
get string network
}
