% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_triselector.R
\name{triselector_module}
\alias{triselector_module}
\title{Utility triselector ui}
\usage{
triselector_module(input, output, session, reactive_x,
  reactive_selector1 = reactive(NULL),
  reactive_selector2 = reactive(NULL),
  reactive_selector3 = reactive(NULL), label = "Group Label:",
  suspendWhenHidden = TRUE)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{reactive_x}{an nx3 matrix}

\item{reactive_selector1}{default value for selector 1}

\item{reactive_selector2}{default value for selector 2}

\item{reactive_selector3}{default value for selector 3}

\item{label}{of the triselector}

\item{suspendWhenHidden}{seuspend when hidden}
}
\description{
Utility triselector ui
}
\examples{
# ### examples
# 
# dat <- readRDS("Dat/exampleEset.RDS")
# library(shiny)
# phenoData <- pData(dat)
# triset <- stringr::str_split_fixed(colnames(phenoData), '\\\\|', n= 3)
# 
# ui <- fluidPage(
#   triselector_ui("tres")
# )
# server <- function(input, output, session) {
#   observe(print(v()))
# }
# 
# shinyApp(ui, server)

# ###
# 
# 
# dat <- readRDS("Dat/exampleEset.RDS")
# library(shiny)
# phenoData <- pData(dat)
# triset <- stringr::str_split_fixed(colnames(phenoData), '\\\\|', n= 3)
# 
# ui <- fluidPage(
#   triselector_ui("tres"),
#   triselector_ui("tres2")
# )
# server <- function(input, output, session) {
#   v1 <- callModule(triselector_module, id = "tres", x = reactive(triset))
#   v2 <- callModule(triselector_module, id = "tres2", x = reactive(triset), 
#                    reactive_selector1 = reactive(v1()$analysis), 
#                    reactive_selector2 = reactive(v1()$subset))
#   observe({
#     print("/////////////////////////")
#     print(v1())
#     })
# }
# 
# shinyApp(ui, server)

}
